+++ vclib-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciProcessTest.kt
@@ -171,7 +171,11 @@ private suspend fun runProcessWithJwtProof(
     requestOptions: WalletService.RequestOptions,
 ): CredentialResponseParameters {
     val token = runProcessGetToken(authorizationService, client, requestOptions)
-    val credentialRequest = client.createCredentialRequestJwt(token, issuer.metadata, requestOptions).getOrThrow()
+    val credentialRequest = client.createCredentialRequestJwt(
+        requestOptions,
+        token.clientNonce,
+        issuer.metadata.credentialIssuer
+    ).getOrThrow()
     return issuer.credential(token.accessToken, credentialRequest).getOrThrow()
 }
 
@@ -182,7 +186,11 @@ private suspend fun runProcessWithCwtProof(
     requestOptions: WalletService.RequestOptions,
 ): CredentialResponseParameters {
     val token = runProcessGetToken(authorizationService, client, requestOptions)
-    val credentialRequest = client.createCredentialRequestCwt(token, issuer.metadata, requestOptions).getOrThrow()
+    val credentialRequest = client.createCredentialRequestCwt(
+        requestOptions = requestOptions,
+        clientNonce = token.clientNonce,
+        credentialIssuer = issuer.metadata.credentialIssuer
+    ).getOrThrow()
     return issuer.credential(token.accessToken, credentialRequest).getOrThrow()
 }
 
@@ -196,7 +204,11 @@ private suspend fun runProcessGetToken(
     authnResponse.shouldBeInstanceOf<AuthenticationResponseResult.Redirect>()
     val code = authnResponse.params.code
     code.shouldNotBeNull()
-    val tokenRequest = client.createTokenRequestParameters(authnResponse.params, requestOptions)
+    val tokenRequest = client.createTokenRequestParameters(
+        requestOptions = requestOptions,
+        code = code,
+        state = authnResponse.params.state!!,
+    )
     val token = authorizationService.token(tokenRequest).getOrThrow()
     return token
 }