+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Release NEXT:
  - Change usage of `OidcUserInfo` in interfaces to `OidcUserInfoExtended`, to also deserialize unknown properties
+ - OID4VCI: `WalletService`: Replace parameters containing whole authentication parameters with single parameters holding `code` and `state`
 
 Release 3.6.1:
  * Update to KMP-Crypto 2.6.0